misc/requirements.txt
numpy
scipy
bagpy
pyyaml
